requirements.txt
Flask>=2.0
flask_cors
flask_jwt_extended
openai
langchain
langchain-openai
langchain-community
langchain-chroma
chroma
pypdf2
pycryptodome
pdfplumber
numpy
pytest
python-dotenv
python-docx
pandas
chromadb
tiktoken
Werkzeug>=2.0
requests
BeautifulSoup4